fix: file upload

+++ src/app/api/beers/[id]/route.ts
@@ -1,8 +1,12 @@
 import { NextRequest, NextResponse } from "next/server";
 import { PrismaClient } from "@/prisma/generated/client";
-import { beerFormSchema } from "@/services/beers/schema";
 import { getServerSession } from "next-auth";
 import { authOptions } from "../../auth/[...nextauth]/route";
+import { beerServerSchema } from "@/services/beers/schema";
+import { notFound } from "next/navigation";
+import path from "path";
+import { unlink } from "fs/promises";
+import { revalidatePath } from "next/cache";
 
 const prisma = new PrismaClient();
 
@@ -19,15 +23,38 @@ export async function PATCH(
 		const body = await req.json();
 
 		// Optional: allow partial updates, or validate full object
-		const result = beerFormSchema.safeParse(body);
+		const result = beerServerSchema.safeParse(body);
 		if (!result.success) {
-			return NextResponse.json({ error: "Invalid input" }, { status: 400 });
+			return NextResponse.json({ error: result.error }, { status: 400 });
+		}
+
+		const beerBeforeUpdate = await prisma.beer.findFirst({ where: { id } });
+		if (!beerBeforeUpdate) {
+			return notFound();
+		}
+
+		if (beerBeforeUpdate.image !== body.image) {
+			const oldImageUrl = beerBeforeUpdate.image;
+			if (oldImageUrl && oldImageUrl.startsWith("/media/")) {
+				const oldFilePath = path.join(process.cwd(), "public", oldImageUrl);
+				try {
+					await unlink(oldFilePath);
+					console.log(`Deleted unused image: ${oldImageUrl}`);
+				} catch (err) {
+					console.warn(
+						`Image file not found or couldn't be deleted: ${oldImageUrl}`
+					);
+					// Don't block the request even if deletion fails
+				}
+			}
 		}
 
 		const beer = await prisma.beer.update({
 			where: { id },
 			data: body,
 		});
+
+		revalidatePath(`dashboard/beers/${id}`, "page");
 
 		return NextResponse.json(beer, { status: 200 });
 	} catch (error: any) {
@@ -52,11 +79,33 @@ export async function DELETE(
 			return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
 		}
 		const id = params.id;
-		const beer = await prisma.beer.delete({
+
+		const beer = await prisma.beer.findUnique({
 			where: { id },
 		});
 
-		return NextResponse.json(beer, { status: 200 });
+		if (!beer) {
+			return NextResponse.json({ error: "Beer not found" }, { status: 404 });
+		}
+
+		if (beer.image) {
+			const imagePath = path.join(process.cwd(), "public", beer.image);
+			try {
+				await unlink(imagePath);
+				console.log(`Deleted image associated with beer: ${beer.image}`);
+			} catch (err) {
+				console.warn(
+					`Image file not found or couldn't be deleted: ${beer.image}`
+				);
+				// Don't block deletion if file is already gone or locked
+			}
+		}
+
+		await prisma.beer.delete({
+			where: { id },
+		});
+
+		return NextResponse.json({ status: 200 });
 	} catch (error: any) {
 		console.error(error);
 		if (
